+++ pkg/operator/object/factory/tests/config_map_test.go
@@ -16,7 +16,7 @@ ingressAdapters:
     action: read
     variant: rest
     route: get
-  steps:
+  tasks:
   - egressAdapter:
       stateless:
         variant: rest