+++ pkg/operator/validation/fields.go
@@ -13,12 +13,12 @@ func ValidateFields(serviceUnitConfig *model.ServiceUnitConfig) ConfigValidation
 
 	for i := range serviceUnitConfig.AdapterConfigs {
 		configValidationError.Extend(validatePrimaryAdapterConfigFields(serviceUnitConfig, &serviceUnitConfig.AdapterConfigs[i]))
-		for _, step := range serviceUnitConfig.AdapterConfigs[i].Tasks {
+		for _, task := range serviceUnitConfig.AdapterConfigs[i].Tasks {
 			// ensure that secondary adapter is defined
-			if step.AdapterConfig == nil {
+			if task.AdapterConfig == nil {
 				continue
 			}
-			configValidationError.Extend(validateSecondaryAdapterConfigFields(*step.AdapterConfig))
+			configValidationError.Extend(validateSecondaryAdapterConfigFields(*task.AdapterConfig))
 		}
 	}
 
@@ -65,7 +65,7 @@ func validatePrimaryAdapterConfigFields(serviceUnitConfig *model.ServiceUnitConf
 	}
 	if primaryAdapterConfig.RepositoryConfig != nil {
 		if len(primaryAdapterConfig.Tasks) > 0 {
-			primaryAdapterConfig.Tasks = []model.TaskSpec{} // makes sure that repository service unit config have no steps defined
+			primaryAdapterConfig.Tasks = []model.TaskSpec{} // makes sure that repository service unit config have no tasks defined
 			logger.Logger.Warnf(
 				"Steps definition found on repository config for %s. These Steps will be ignored.\n",
 				serviceUnitConfig.Name,
@@ -76,12 +76,12 @@ func validatePrimaryAdapterConfigFields(serviceUnitConfig *model.ServiceUnitConf
 			serviceUnitConfig.Name,
 		)
 	}
-	var stepFieldErrors []InvalidStepFieldValueError
-	for _, step := range primaryAdapterConfig.Tasks {
-		errs := ValidateStepFields(step, primaryAdapterConfig, serviceUnitConfig.Name)
-		stepFieldErrors = append(stepFieldErrors, errs...)
+	var taskFieldErrors []InvalidStepFieldValueError
+	for _, task := range primaryAdapterConfig.Tasks {
+		errs := ValidateStepFields(task, primaryAdapterConfig, serviceUnitConfig.Name)
+		taskFieldErrors = append(taskFieldErrors, errs...)
 	}
-	return ConfigValidationError{AdapterFieldErrors: adapterFieldErrors, StepFieldErrors: stepFieldErrors}
+	return ConfigValidationError{AdapterFieldErrors: adapterFieldErrors, StepFieldErrors: taskFieldErrors}
 }
 
 // Validate the fields of the secondary adapter configuration