+++ docs/changelog.md
@@ -5,9 +5,9 @@
 - `.ingressAdapters.0.broker` -> `.adapters.0.consumer`
 - `.ingressAdapters.0.stateless` -> `.adapters.0.server`
 - `.ingressAdapters.0.stateful` -> `.adapters.0.repository`
-- `.ingressAdapters.0.steps.0.egressAdapter.broker` -> `.ingressAdapters.0.steps.0.adapter.producer`
-- `.ingressAdapters.0.steps.0.egressAdapter.stateless` -> `.ingressAdapters.0.steps.0.adapter.invocation`
-- `.ingressAdapters.0.steps.0.egressAdapter.stateful` -> `.ingressAdapters.0.steps.0.adapter.repository`
+- `.ingressAdapters.0.tasks.0.egressAdapter.broker` -> `.ingressAdapters.0.tasks.0.adapter.producer`
+- `.ingressAdapters.0.tasks.0.egressAdapter.stateless` -> `.ingressAdapters.0.tasks.0.adapter.invocation`
+- `.ingressAdapters.0.tasks.0.egressAdapter.stateful` -> `.ingressAdapters.0.tasks.0.adapter.repository`
 
 ### New Fields
 - `deployment`: config for deployment